requirements.txt
Flask
azure-identity
azure-keyvault-secrets
